+++ shub/apps/base/views/search.py
@@ -9,14 +9,22 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 from django.shortcuts import render
+from ratelimit.decorators import ratelimit
+
+from shub.settings import (
+    VIEW_RATE_LIMIT as rl_rate, 
+    VIEW_RATE_LIMIT_BLOCK as rl_block
+)
+
 
 # Search Pages #################################################################
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def search_view(request):
     context = {'active':'share'}
     return render(request, 'search/search.html', context)
 
-
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def search_query(request, query=None):
     '''query is a post, and results returned immediately'''
     from shub.apps.main.query import collection_query
@@ -27,7 +35,7 @@ def search_query(request, query=None):
  
     return render(request, 'search/search_single_page.html', context)
 
-
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def container_search(request):
     '''container_search is the ajax driver to show results for a container search.
     by default we search container collection name.