+++ shub/apps/main/views/tags.py
@@ -9,6 +9,10 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 from shub.apps.main.models import Container
+from shub.settings import (
+    VIEW_RATE_LIMIT as rl_rate, 
+    VIEW_RATE_LIMIT_BLOCK as rl_block
+)
 
 from taggit.models import Tag
 
@@ -21,6 +25,7 @@ from django.http import JsonResponse
 from django.http.response import Http404
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from ratelimit.decorators import ratelimit
 
 from .containers import get_container
 
@@ -44,15 +49,16 @@ def get_tag(name=None, tid=None):
 # TAGS #########################################################################
 ################################################################################
 
-
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def all_tags(request):
     tags = Tag.objects.all()
     context = {"tags":tags}
     return render(request, 'tags/all_tags.html', context)
 
 
-# View containers for a tag
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def view_tag(request, tid):
+    '''view containers for a tag'''
     try:
         tag = Tag.objects.get(id=tid)
     except:
@@ -70,6 +76,7 @@ def view_tag(request, tid):
 # COLLECTION TAG MANAGEMENT
 ################################################################################
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def add_tag(request, cid):
     '''manually add a tag to the collection
@@ -88,7 +95,7 @@ def add_tag(request, cid):
     return JsonResponse({"message":message})
 
 
-
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def remove_tag(request, cid):
     '''remove a tag from a collection