+++ shub/apps/users/views/teams.py
@@ -17,6 +17,7 @@ from django.http import (
     JsonResponse
 )
 
+from ratelimit.decorators import ratelimit
 from shub.apps.users.forms import TeamForm
 from shub.apps.users.models import (
     Team,
@@ -27,6 +28,10 @@ from shub.apps.users.permissions import (
     is_invite_valid 
 )
 from shub.apps.users.utils import get_user
+from shub.settings import (
+    VIEW_RATE_LIMIT as rl_rate, 
+    VIEW_RATE_LIMIT_BLOCK as rl_block
+)
 
 import uuid
 
@@ -54,6 +59,7 @@ def get_team(tid):
 ################################################################################
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def edit_team(request, tid=None):
     '''edit_team is the view to edit an existing team, or create a new team.
@@ -104,6 +110,7 @@ def edit_team(request, tid=None):
     return redirect("teams")
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 def view_teams(request):
     '''view all teams (log in not required)
     :parma tid: the team id to edit or create. If none, indicates a new team
@@ -119,6 +126,7 @@ def view_teams(request):
     return render(request, "teams/all_teams.html", context)
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def view_team(request, tid, code=None):
     '''view the details about a team
@@ -145,6 +153,7 @@ def view_team(request, tid, code=None):
 ################################################################################
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def join_team(request, tid, code=None):
     '''add a user to a new team. If the team is open, he/she can join without
@@ -187,6 +196,7 @@ def join_team(request, tid, code=None):
 ################################################################################
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def leave_team(request, tid):
     '''leave team is the view for a user to leave his or her team. A user
@@ -234,6 +244,7 @@ def _remove_owner(team, user):
     return team
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def remove_member(request, tid, uid):
     '''remove a member from a team.
@@ -255,6 +266,7 @@ def remove_member(request, tid, uid):
     return JsonResponse({"message":message})
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def remove_owner(request, tid, uid):
     '''remove a member from a team.
@@ -276,6 +288,7 @@ def remove_owner(request, tid, uid):
     return JsonResponse({"message":message})
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def add_owner(request, tid, uid):
     '''promote a user to be owner of a team
@@ -298,6 +311,7 @@ def add_owner(request, tid, uid):
     return JsonResponse({"message":message})
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def delete_team(request, tid):
     '''delete a team entirely, must be an owner
@@ -318,6 +332,7 @@ def delete_team(request, tid):
     return redirect('teams')
 
 
+@ratelimit(key='ip', rate=rl_rate, block=rl_block)
 @login_required
 def generate_team_invite(request, tid):
     '''generate an invitation for a user, return to view.